Re factored maven repository structure into core , test common and junit, this will allow us to include support for TestNG in the feature.
