+++ filemock-core/src/test/resources/META-INF/mock/sample_pojo_more.json
