--- src/test/java/test/kvlabs/filemock/model/jackson1/SubObject.java
@@ -1,53 +1,0 @@
-
-package test.kvlabs.filemock.model.jackson1;
-
-import org.codehaus.jackson.annotate.JsonProperty;
-import org.codehaus.jackson.annotate.JsonPropertyOrder;
-import org.codehaus.jackson.map.annotate.JsonSerialize;
-
-@JsonSerialize(include = JsonSerialize.Inclusion.NON_NULL)
-@JsonPropertyOrder({
-    "a",
-    "c",
-    "e"
-})
-public class SubObject {
-
-    @JsonProperty("a")
-    private String a;
-    @JsonProperty("c")
-    private String c;
-    @JsonProperty("e")
-    private String e;
-
-    @JsonProperty("a")
-    public String getA() {
-        return a;
-    }
-
-    @JsonProperty("a")
-    public void setA(String a) {
-        this.a = a;
-    }
-
-    @JsonProperty("c")
-    public String getC() {
-        return c;
-    }
-
-    @JsonProperty("c")
-    public void setC(String c) {
-        this.c = c;
-    }
-
-    @JsonProperty("e")
-    public String getE() {
-        return e;
-    }
-
-    @JsonProperty("e")
-    public void setE(String e) {
-        this.e = e;
-    }
-
-}
