--- src/test/java/test/kvlabs/filemock/model/jackson2/SubObject.java
@@ -1,52 +1,0 @@
-package test.kvlabs.filemock.model.jackson2;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonPropertyOrder({
-    "a",
-    "c",
-    "e"
-})
-public class SubObject {
-
-    @JsonProperty("a")
-    private String a;
-    @JsonProperty("c")
-    private String c;
-    @JsonProperty("e")
-    private String e;
-
-    @JsonProperty("a")
-    public String getA() {
-        return a;
-    }
-
-    @JsonProperty("a")
-    public void setA(String a) {
-        this.a = a;
-    }
-
-    @JsonProperty("c")
-    public String getC() {
-        return c;
-    }
-
-    @JsonProperty("c")
-    public void setC(String c) {
-        this.c = c;
-    }
-
-    @JsonProperty("e")
-    public String getE() {
-        return e;
-    }
-
-    @JsonProperty("e")
-    public void setE(String e) {
-        this.e = e;
-    }
-
-}
